wip: pongEntities.js and preparing for remote ball updates
